+++ Runtime/ListenerComponents/FloatEventVarListener.cs
@@ -3,6 +3,5 @@ namespace BardicBytes.EventVars
 {
     public class FloatEventVarListener : EventVarListener<float>
     {
-        protected override void HandleTypedEventRaised(float data) => base.HandleTypedEventRaised(data);
     }
 }