+++ Runtime/ListenerComponents/GameObjectEventVarListener.cs
@@ -6,6 +6,5 @@ namespace BardicBytes.EventVars
 {
     public class GameObjectEventVarListener : EventVarListener<GameObject>
     {
-        protected override void HandleTypedEventRaised(GameObject data) => base.HandleTypedEventRaised(data);
     }
 }