+++ Samples~/Demo/Scripts/DemoButton.cs
@@ -1,3 +1,5 @@
+//alex@bardicbytes.com
+
 using BardicBytes.EventVars;
 using TMPro;
 using UnityEngine;