+++ Samples~/TypedEventVars/RigidbodyEventVarListener.cs
@@ -1,0 +1,8 @@
+//alex@bardicbytes.com
+
+using BardicBytes.EventVars;
+using UnityEngine;
+
+public class RigidbodyEventVarListener : EventVarListener<Rigidbody>
+{
+}